+++ src/main/java/gov/nasa/pds/registry/common/connection/es/RequestConfigCB.java
@@ -1,4 +1,4 @@
-package gov.nasa.pds.registry.common.es.client;
+package gov.nasa.pds.registry.common.connection.es;
 
 
 import org.apache.http.client.config.RequestConfig;
