+++ src/main/java/gov/nasa/pds/registry/common/connection/es/SearchResponseParser.java
@@ -1,4 +1,4 @@
-package gov.nasa.pds.registry.common.es.client;
+package gov.nasa.pds.registry.common.connection.es;
 
 import java.io.InputStream;
 import java.io.InputStreamReader;
