+++ src/test/java/dao/TestProductDao.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 import org.elasticsearch.client.RestClient;
 import gov.nasa.pds.registry.common.EstablishConnectionFactory;
-import gov.nasa.pds.registry.common.es.client.EsClientFactory;
+import gov.nasa.pds.registry.common.connection.es.EsClientFactory;
 import gov.nasa.pds.registry.common.es.dao.ProductDao;
 import gov.nasa.pds.registry.common.util.CloseUtils;
 import tt.TestLogConfigurator;
