+++ src/test/java/tt/TestRestClient.java
@@ -2,7 +2,7 @@ package tt;
 
 import org.elasticsearch.client.RestClient;
 import gov.nasa.pds.registry.common.EstablishConnectionFactory;
-import gov.nasa.pds.registry.common.es.client.EsClientFactory;
+import gov.nasa.pds.registry.common.connection.es.EsClientFactory;
 
 
 public class TestRestClient
